bom_raw_products calculation

+++ userService/src/main/java/com/nector/userservice/bom/dto/BomProductionResponseDto.java
@@ -36,6 +36,6 @@ public class BomProductionResponseDto {
         private BigDecimal quantityRequired;
         private BigDecimal quantityAvailable;
         private String unit;
-        private Integer newStock;
+        private BigDecimal newStock;
     }
 }
